refactor: small changes to importAndSaveCommitments for consistency
