+++ libs/nightfall/isCommitmentFromMnemonic.ts
